Move the BasePipeline class to a new `aboutcode.pipeline` module #1351
