+++ scanpipe/pipelines/__init__.py
@@ -26,17 +26,12 @@ import traceback
 from contextlib import contextmanager
 from functools import wraps
 from pathlib import Path
-from pydoc import getdoc
-from pydoc import splitdoc
-from timeit import default_timer as timer
-
-from django.utils import timezone
 
 import bleach
 from markdown_it import MarkdownIt
 from pyinstrument import Profiler
 
-from scanpipe import humanize_time
+from aboutcode.pipeline import BasePipeline
 
 logger = logging.getLogger(__name__)
 
@@ -55,19 +50,6 @@ class InputFilesError(Exception):
         return message
 
 
-def group(*groups):
-    """Mark a function as part of a particular group."""
-
-    def decorator(obj):
-        if hasattr(obj, "groups"):
-            obj.groups = obj.groups.union(groups)
-        else:
-            setattr(obj, "groups", set(groups))
-        return obj
-
-    return decorator
-
-
 def convert_markdown_to_html(markdown_text):
     """Convert Markdown text to sanitized HTML."""
     # Using the "js-default" for safety.
@@ -77,13 +59,11 @@ def convert_markdown_to_html(markdown_text):
     return sanitized_html
 
 
-class BasePipeline:
-    """Base class for all pipelines."""
+class Pipeline(BasePipeline):
+    """Main class for all Project pipelines including common step methods."""
 
     # Flag specifying whether to download missing inputs as an initial step.
     download_inputs = True
-    # Flag indicating if the Pipeline is an add-on, meaning it cannot be run first.
-    is_addon = False
     # Optional URL that targets a view of the results relative to this Pipeline.
     # This URL may contain dictionary-style string formatting, which will be
     # interpolated against the project's field attributes.
@@ -91,151 +71,24 @@ class BasePipeline:
     # to target the Package list view with an active filtering.
     results_url = ""
 
-    def __init__(self, run):
-        """Load the Run and Project instances."""
-        self.run = run
-        self.project = run.project
-        self.pipeline_name = run.pipeline_name
-        self.env = self.project.get_env()
-
     @classmethod
-    def steps(cls):
-        raise NotImplementedError
-
-    @classmethod
-    def get_steps(cls, groups=None):
-        """
-        Return the list of steps defined in the ``steps`` class method.
-
-        If the optional ``groups`` parameter is provided, only include steps labeled
-        with groups that intersect with the provided list. If a step has no groups or
-        if ``groups`` is not specified, include the step in the result.
-        """
-        if not callable(cls.steps):
-            raise TypeError("Use a ``steps(cls)`` classmethod to declare the steps.")
-
-        steps = cls.steps()
-
-        if groups is not None:
-            steps = tuple(
-                step
-                for step in steps
-                if not getattr(step, "groups", [])
-                or set(getattr(step, "groups")).intersection(groups)
-            )
-
-        return steps
-
-    @classmethod
-    def get_doc(cls):
-        """Get the doc string of this pipeline."""
-        return getdoc(cls)
-
-    @classmethod
-    def get_graph(cls):
-        """Return a graph of steps."""
-        return [
-            {
-                "name": step.__name__,
-                "doc": getdoc(step),
-                "groups": getattr(step, "groups", []),
-            }
-            for step in cls.get_steps()
-        ]
+    def get_initial_steps(cls):
+        """Add the ``download_inputs`` step as an initial step if enabled."""
+        if cls.download_inputs:
+            return (cls.download_missing_inputs,)
 
     @classmethod
     def get_info(cls, as_html=False):
-        """Get a dictionary of combined information data about this pipeline."""
-        summary, description = splitdoc(cls.get_doc())
-        steps = cls.get_graph()
+        """Add the option to render the values as HTML."""
+        info = super().get_info()
 
         if as_html:
-            summary = convert_markdown_to_html(summary)
-            description = convert_markdown_to_html(description)
-            for step in steps:
+            info["summary"] = convert_markdown_to_html(info["summary"])
+            info["description"] = convert_markdown_to_html(info["description"])
+            for step in info["steps"]:
                 step["doc"] = convert_markdown_to_html(step["doc"])
 
-        return {
-            "summary": summary,
-            "description": description,
-            "steps": steps,
-            "available_groups": cls.get_available_groups(),
-        }
-
-    @classmethod
-    def get_summary(cls):
-        """Get the doc string summary."""
-        return cls.get_info()["summary"]
-
-    @classmethod
-    def get_available_groups(cls):
-        return sorted(
-            set(
-                group_name
-                for step in cls.get_steps()
-                for group_name in getattr(step, "groups", [])
-            )
-        )
-
-    def log(self, message):
-        """Log the given `message` to the current module logger and Run instance."""
-        now_as_localtime = timezone.localtime(timezone.now())
-        timestamp = now_as_localtime.strftime("%Y-%m-%d %H:%M:%S.%f")[:-4]
-        message = f"{timestamp} {message}"
-        logger.info(message)
-        self.run.append_to_log(message)
-
-    @staticmethod
-    def output_from_exception(exception):
-        """Return a formatted error message including the traceback."""
-        output = f"{exception}\n\n"
-
-        if exception.__cause__ and str(exception.__cause__) != str(exception):
-            output += f"Cause: {exception.__cause__}\n\n"
-
-        traceback_formatted = "".join(traceback.format_tb(exception.__traceback__))
-        output += f"Traceback:\n{traceback_formatted}"
-
-        return output
-
-    def execute(self):
-        """Execute each steps in the order defined on this pipeline class."""
-        self.log(f"Pipeline [{self.pipeline_name}] starting")
-
-        steps = self.get_steps(groups=self.run.selected_groups)
-        selected_steps = self.run.selected_steps
-
-        if self.download_inputs:
-            steps = (self.__class__.download_missing_inputs,) + steps
-
-        steps_count = len(steps)
-        pipeline_start_time = timer()
-
-        for current_index, step in enumerate(steps, start=1):
-            step_name = step.__name__
-
-            if selected_steps and step_name not in selected_steps:
-                self.log(f"Step [{step_name}] skipped")
-                continue
-
-            self.run.set_current_step(f"{current_index}/{steps_count} {step_name}")
-            self.log(f"Step [{step_name}] starting")
-            step_start_time = timer()
-
-            try:
-                step(self)
-            except Exception as exception:
-                self.log("Pipeline failed")
-                return 1, self.output_from_exception(exception)
-
-            step_run_time = timer() - step_start_time
-            self.log(f"Step [{step_name}] completed in {humanize_time(step_run_time)}")
-
-        self.run.set_current_step("")  # Reset the `current_step` field on completion
-        pipeline_run_time = timer() - pipeline_start_time
-        self.log(f"Pipeline completed in {humanize_time(pipeline_run_time)}")
-
-        return 0, ""
+        return info
 
     def download_missing_inputs(self):
         """
@@ -295,10 +148,6 @@ class BasePipeline:
             yield
         except exceptions as error:
             self.add_error(exception=error, **kwargs)
-
-
-class Pipeline(BasePipeline):
-    """Main class for all pipelines including common step methods."""
 
     def flag_empty_files(self):
         """Flag empty files."""
